aiohttp/requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0